API/requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.32
PyMySQL==1.0.3
Werkzeug==2.2.3
zipp==3.15.0
boto3==1.26.142
beautifulsoup4==4.12.0
bs4==0.0.1
requests==2.28.2
requests-html==0.10.0
requests-toolbelt==0.10.1
manga109api==0.3.1
Pillow==10.0.0